+++ inst/examples/ex-assign.R
@@ -1,0 +1,11 @@
+X <- data.frame(
+  x = 1:5,
+  y = 6:10,
+  z = LETTERS[1:5]
+)
+
+## Assign column to row names
+(Y <- assign_rownames(X, 3))
+
+## Append row names to data.frame
+(Z <- append_rownames(Y))