+++ inst/examples/ex-reshape.R
@@ -1,11 +1,5 @@
-X <- data.frame(
-  x = 1:5,
-  y = 6:10,
-  z = LETTERS[1:5]
-)
+## Create a matrix
+A <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
 
-## Assign column to row names
-(Y <- assign_rownames(X, 3))
-
-## Append row names to data.frame
-(Z <- append_rownames(Y))
+## Transform to long data.frame
+head(to_long(A))